refactor: 为MTP Runtime建立子包，作为MTP执行引擎
