+++ src/hivememory/agent_runtime/mtp/runtime.py
@@ -118,7 +118,7 @@ class KoakumaRuntime:
 
         # 初始化内核工具注册表 KERNEL_REGISTRY (Section 4.2.1)
         # 硬编码的 sys_ 工具集，随系统启动加载，Zero Latency
-        from hivememory.agent_runtime.syscalls import build_kernel_registry
+        from hivememory.agent_runtime.mtp.syscalls import build_kernel_registry
         self._kernel_registry = build_kernel_registry(
             python_repl_timeout=self._config.python_repl_timeout_seconds,
             workspace_path=self._config.workspace_path,
@@ -907,7 +907,7 @@ class KoakumaRuntime:
         Returns:
             MTPResponse: 执行结果
         """
-        from hivememory.agent_runtime.syscalls import execute_sandboxed
+        from hivememory.agent_runtime.mtp.syscalls import execute_sandboxed
 
         result = execute_sandboxed(
             code,
