+++ src/hivememory/agent_runtime/mtp/syscalls/clock.py
