+++ src/hivememory/agent_runtime/mtp/syscalls/file_io.py
