+++ src/hivememory/agent_runtime/mtp/syscalls/repl.py
