+++ src/hivememory/agent_runtime/mtp/syscalls/types.py
