+++ tests/e2e/pipeline/test_agent_permission_e2e.py
@@ -15,7 +15,7 @@ from uuid import uuid4
 
 from hivememory.core.models import AgentProfile, MemoryAtom, MetaData, IndexLayer, PayloadLayer, MemoryType, OMNI_DOLL_PROFILE
 from hivememory.patchouli.runtime.core import PatchouliRuntime
-from hivememory.agent_runtime.koakuma import KoakumaRuntime
+from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
 from hivememory.agent_runtime.models import MTPExecutionContext
 from hivememory.core.mtp.exceptions import PermissionDeniedError
 from hivememory.core.mtp import MTPCommand, MTPVerb
