+++ tests/unit/alice/runtime/syscalls/test_live_file_io.py
@@ -7,7 +7,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from hivememory.system.config import KoakumaConfig
-from hivememory.agent_runtime.koakuma import KoakumaRuntime
+from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
 
 from .live_support import (
     MTPLoopRunner,
