+++ tests/unit/alice/runtime/syscalls/test_repl.py
@@ -1,4 +1,4 @@
-from hivememory.agent_runtime.syscalls.repl import sys_python_repl
+from hivememory.agent_runtime.mtp.syscalls.repl import sys_python_repl
 
 
 class TestSysPythonRepl:
