+++ tests/unit/patchouli/mtp/test_mtp_scenario_live.py
@@ -28,7 +28,7 @@ from hivememory.core.models import (
 from hivememory.system.config import KoakumaConfig
 from hivememory.core.mtp.models import MTP_LEFT_DELIMITER
 from hivememory.core.protocol.models import RetrievalResponse
-from hivememory.agent_runtime.koakuma import KoakumaRuntime
+from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
 from hivememory.agent_runtime.models import MTPExecutionContext
 from hivememory.prompts.mtp import MTPPromptBuilder
