+++ tests/unit/patchouli/mtp/test_search_chain.py
@@ -25,7 +25,7 @@ from hivememory.core.models import (
 )
 from hivememory.engines.retrieval.models import QueryFilters
 from hivememory.core.protocol.models import RetrievalResponse
-from hivememory.agent_runtime.koakuma import KoakumaRuntime
+from hivememory.agent_runtime.mtp.runtime import KoakumaRuntime
 from hivememory.agent_runtime.models import MTPExecutionContext
 from hivememory.system.config import KoakumaConfig
 from hivememory.core.mtp import MTPResponseStatus
